+++ lib/hooks/useBlockRenderer.d.ts
@@ -1,0 +1,12 @@
+declare type UseRetrieveBlockRenderer = {
+    name?: string;
+    context?: 'view' | 'edit';
+    attributes?: object;
+    post_id?: number;
+};
+export declare const useRetrieveBlockRenderer: (options: UseRetrieveBlockRenderer) => {
+    data: object[];
+    loading: boolean;
+    error: string | object | null;
+};
+export {};