+++ lib/hooks/useTypes.d.ts
@@ -1,15 +1,16 @@
 declare type UseTypes = {
-    context: 'view' | 'embed' | 'edit';
+    context?: 'view' | 'embed' | 'edit';
 };
-declare type UseRetriveType = {
-    context: 'view' | 'embed' | 'edit';
+declare type UseRetrieveType = {
+    type?: string;
+    context?: 'view' | 'embed' | 'edit';
 };
 export declare const useTypes: (options: UseTypes) => {
     data: object[];
     loading: boolean;
     error: string | object | null;
 };
-export declare const useRetriveType: (options: UseRetriveType) => {
+export declare const useRetrieveType: (options: UseRetrieveType) => {
     data: object[];
     loading: boolean;
     error: string | object | null;